+++ saleor/account/i18n.py
@@ -3,13 +3,11 @@ from collections import defaultdict
 import i18naddress
 from django import forms
 from django.forms.forms import BoundField
-from django.utils.translation import pgettext_lazy, ugettext_lazy as _
+from django.utils.translation import pgettext_lazy
 from django_countries import countries
-from phonenumber_field.phonenumber import PhoneNumber
-from phonenumbers import NumberParseException
-from phonenumbers.phonenumberutil import is_possible_number
 
 from .models import Address
+from .validators import validate_possible_number
 from .widgets import DatalistTextWidget, PhonePrefixWidget
 
 COUNTRY_FORMS = {}
@@ -145,24 +143,10 @@ class AddressForm(forms.ModelForm):
         country = data.get("country")
         if phone:
             try:
-                data["phone"] = clean_phone_for_country(phone, country)
+                data["phone"] = validate_possible_number(phone, country)
             except forms.ValidationError as error:
                 self.add_error("phone", error)
         return data
-
-
-def clean_phone_for_country(phone, country):
-    error = _("The phone number entered is not valid.")
-    error_code = "invalid_phone_number"
-    if phone:
-        try:
-            phone = PhoneNumber.from_string(phone, country)
-        except NumberParseException:
-            raise forms.ValidationError(error, code=error_code)
-        else:
-            if not is_possible_number(phone):
-                raise forms.ValidationError(error, code=error_code)
-    return phone
 
 
 class CountryAwareAddressForm(AddressForm):